backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
anthropic==0.34.2
sentence-transformers==3.1.1
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
